Add tests for fdbcli quota commands

+++ fdbclient/TagThrottle.actor.cpp
@@ -150,8 +150,8 @@ ThrottleApi::TagQuotaValue ThrottleApi::TagQuotaValue::fromValue(ValueRef value)
 	}
 	TagQuotaValue result;
 	try {
-		result.reservedQuota = tuple.getDouble(0);
-		result.totalQuota = tuple.getDouble(1);
+		result.reservedQuota = tuple.getInt(0);
+		result.totalQuota = tuple.getInt(1);
 	} catch (Error& e) {
 		TraceEvent(SevWarnAlways, "TagQuotaValueFailedToDeserialize").error(e);
 		throw invalid_throttle_quota_value();
